+++ checkpoint/orbax/checkpoint/pytree_checkpoint_handler.py
@@ -862,7 +862,7 @@ class PyTreeCheckpointHandler(async_checkpoint_handler.AsyncCheckpointHandler):
                 )
             )[0]
         )
-      utils.sync_global_devices(
+      utils.sync_global_processes(
           'PyTreeCheckpointHandler:create_param_save_dirs'
       )
 
@@ -928,7 +928,7 @@ class PyTreeCheckpointHandler(async_checkpoint_handler.AsyncCheckpointHandler):
           f.result()  # Block on result.
 
     asyncio.run(async_save(directory, *args, **kwargs))
-    utils.sync_global_devices('PyTreeCheckpointHandler:save')
+    utils.sync_global_processes('PyTreeCheckpointHandler:save')
 
   async def _maybe_deserialize(
       self, structure: PyTree, param_infos: PyTree, restore_args: PyTree
@@ -1140,7 +1140,7 @@ class PyTreeCheckpointHandler(async_checkpoint_handler.AsyncCheckpointHandler):
           transforms,
           transforms_default_to_original,
       )
-    utils.sync_global_devices('PyTreeCheckpointHandler:restore')
+    utils.sync_global_processes('PyTreeCheckpointHandler:restore')
 
     if logging.level_debug():
       logging.debug('param_infos: %s', param_infos)