+++ checkpoint/orbax/checkpoint/standard_checkpoint_handler_test_utils.py
@@ -45,7 +45,7 @@ class StandardCheckpointHandler(
     super().save(directory, *args, **kwargs)
     if jax.process_index() == 0:
       self.finalize(directory)
-    utils.sync_global_devices('StandardCheckpointHandler:finalize')
+    utils.sync_global_processes('StandardCheckpointHandler:finalize')
 
 
 # Not in common util because we need to eliminate OSS dependency on flax.
@@ -92,10 +92,10 @@ class StandardCheckpointHandlerTestBase:
       )
       test_utils.set_tensorstore_driver_for_test()
 
-      utils.sync_global_devices('StandardCheckpointHandler:setup_complete')
+      utils.sync_global_processes('StandardCheckpointHandler:setup_complete')
 
     def tearDown(self):
-      utils.sync_global_devices('StandardCheckpointHandler:tests_complete')
+      utils.sync_global_processes('StandardCheckpointHandler:tests_complete')
       self.handler.close()
       super().tearDown()